Support Native Facter
To support Native Facter, `facter` has been moved into its own
Bundler group. This makes it a little easier to control whether or not
Classic Facter is installed. This also requires an update to config.ru
to support loading the group, and a change to README detailing the
environment variables that are used to configure usage for Classic and
Native Facter.